+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Add userGroupLinks as a field in UserService
 
+* Remove TypeName.ARRAY and use TypeName.COLLECTION instead
+
 ## 0.7.5
 
 * New TaskService.java base class to reduce duplicate code when